trying to put Welcome msg in center of navbar"

+++ src/screens/Signup.js
@@ -1,70 +1,48 @@
-import React, { useState } from 'react'
-import { useNavigate, Link } from 'react-router-dom'
+import React, { useState } from 'react';
+import { useNavigate, Link } from 'react-router-dom';
 import Navbar from '../components/Navbar';
-import './Signup.css'
+import './Signup.css';
+
 export default function Signup() {
-  const [credentials, setCredentials] = useState({ name: "", email: "", password: "", geolocation: "" })
-  // let [address, setAddress] = useState("");
-  let navigate = useNavigate()
+  const [credentials, setCredentials] = useState({ name: "", email: "", password: "", confirmPassword: "", geolocation: "" });
+  const [passwordError, setPasswordError] = useState("");
 
-  // const handleClick = async (e) => {
-  //   e.preventDefault();
-  //   let navLocation = () => {
-  //     return new Promise((res, rej) => {
-  //       navigator.geolocation.getCurrentPosition(res, rej);
-  //     });
-  //   }
-  //   let latlong = await navLocation().then(res => {
-  //     let latitude = res.coords.latitude;
-  //     let longitude = res.coords.longitude;
-  //     return [latitude, longitude]
-  //   })
-  //   // console.log(latlong)
-  //   let [lat, long] = latlong
-  //   console.log(lat, long)
-  //   const response = await fetch("http://localhost:5000/api/auth/getlocation", {
-  //     method: 'POST',
-  //     headers: {
-  //       'Content-Type': 'application/json'
-  //     },
-  //     body: JSON.stringify({ latlong: { lat, long } })
-
-  //   });
-  //   const { location } = await response.json()
-  //   console.log(location);
-  //   setAddress(location);
-  //   setCredentials({ ...credentials, [e.target.name]: location })
-  // }
+  const navigate = useNavigate();
 
   const handleSubmit = async (e) => {
     e.preventDefault();
+
+    // Validate passwords
+    if (credentials.password !== credentials.confirmPassword) {
+      setPasswordError("Passwords do not match");
+      return;
+    }
+
+    // Clear previous password error
+    setPasswordError("");
+
     const response = await fetch("http://localhost:5000/api/auth/createuser", {
-      // credentials: 'include',
-      // Origin:"http://localhost:3000/login",
       method: 'POST',
       headers: {
         'Content-Type': 'application/json'
       },
       body: JSON.stringify({ name: credentials.name, email: credentials.email, password: credentials.password, location: credentials.geolocation })
-
     });
-    const json = await response.json()
-    console.log(json);
-    if (json.success) {
-      //save the auth toke to local storage and redirect
-      localStorage.setItem('token', json.authToken)
-      localStorage.setItem('userName', credentials.name)
-      navigate("/login")
 
+    const json = await response.json();
+
+    if (json.success) {
+      localStorage.setItem('token', json.authToken);
+      localStorage.setItem('userName', credentials.name);
+      navigate("/login");
+    } else {
+      alert("Enter Valid Credentials");
     }
-    else {
-      alert("Enter Valid Credentials")
-    }
-  }
+  };
 
   const onChange = (e) => {
-    setCredentials({ ...credentials, [e.target.name]: e.target.value })
-  }
+    setCredentials({ ...credentials, [e.target.name]: e.target.value });
+  };
 
   return (
     <div style={{ backgroundImage: 'url("https://cdn.vox-cdn.com/thumbor/8zX9C04GCHA9HvJR7PjsJneqLUg=/0x0:2600x1733/1200x800/filters:focal(1092x659:1508x1075)/cdn.vox-cdn.com/uploads/chorus_image/image/66554227/shutterstock_616076909.0.jpg")', backgroundSize: 'cover', height: '100vh' }}>
@@ -72,26 +50,31 @@ export default function Signup() {
         <Navbar />
       </div>
       <br /> <br /> <br /> <br />
-      <div className='container' >
+      <div className='container'>
         <form className='w-50 m-auto mt-5 border bg-dark border-success rounded' onSubmit={handleSubmit}>
           <div className="m-3">
             <label htmlFor="name" className="form-label">Name</label>
-            <input type="text" className="form-control" name='name' value={credentials.name} onChange={onChange} aria-describedby="emailHelp" />
+            <input type="text" className="form-control" name='name' value={credentials.name} onChange={onChange} />
           </div>
           <div className="m-3">
             <label htmlFor="email" className="form-label">Email address</label>
-            <input type="email" className="form-control" name='email' value={credentials.email} onChange={onChange} aria-describedby="emailHelp" />
+            <input type="email" className="form-control" name='email' value={credentials.email} onChange={onChange} />
           </div>
           <div className="m-3">
-            <label htmlFor="exampleInputPassword1" className="form-label">Password</label>
-            <input type="password" className="form-control" value={credentials.password} onChange={onChange} name='password' />
+            <label htmlFor="password" className="form-label">Password</label>
+            <input type="password" className="form-control" name='password' value={credentials.password} onChange={onChange} />
           </div>
+          <div className="m-3">
+            <label htmlFor="confirmPassword" className="form-label">Confirm Password</label>
+            <input type="password" className="form-control" name='confirmPassword' value={credentials.confirmPassword} onChange={onChange} />
+          </div>
+          {passwordError && <div className="alert alert-danger">{passwordError}</div>}
           <div className='submit-btn'>
-            <button type="submit" className="btn btn-success">Submit</button>
+            <button type="submit" className="btn btn-success">Register</button>
             <Link to="/login" className="mx-1 btn btn-danger">Already a user</Link>
           </div>
         </form>
       </div>
     </div>
-  )
+  );
 }
